fix/ dockfile & dir
