+++ cogs/__init__.py
@@ -6,14 +6,5 @@ if TYPE_CHECKING:
     from core import LIPOIC
 
 
-def loadCommands(bot: 'LIPOIC'):
-    bot.load_extension(".commands.__init__", package=__package__)
-
-
-def loadEvents(bot: 'LIPOIC'):
-    bot.load_extension(".events.__init__", package=__package__)
-
-
-def loadAll(bot: 'LIPOIC'):
-    loadCommands(bot)
-    loadEvents(bot)
+def setup(bot: 'LIPOIC'):
+    bot.load_cog_dir(__package__, __file__, deep=True)
